cancellato alcune classi e scritte nuove e poi ho implementato il metodo aggiungi allievo e funziona tutto bisogna metterci il controllo degli errori ma questo poi si fa

+++ src/main/java/it/uniroma3/spring/repository/AziendaRepository.java
@@ -6,13 +6,13 @@ import org.springframework.data.repository.CrudRepository;
 
 import it.uniroma3.spring.model.Azienda;
 
-public interface AziendaRepository extends CrudRepository<Azienda, Long>{
+public interface AziendaRepository extends CrudRepository<Azienda, Integer>{
 
 	public Azienda findByNome(String nome);
 
 	public List<Azienda>findAll();
 
-	public Azienda findById(Long id);
+	public Azienda findById(int id);
